-Werror=sign-compare

+++ include/block.h
@@ -31,7 +31,7 @@ void block_free(block_t *block);
 int block_append(block_thread_state_t *block_thread_state, block_t *block, index_t *index, const char *data, size_t size);
 int block_flush(block_thread_state_t *block_thread_state, block_t *block, index_t *index, char* ref, int force_indirection);
 
-int block_ref_length(const char *ref);
+size_t block_ref_length(const char *ref);
 
 int block_setup(block_t *block, const char *ref, size_t ref_len);
 ssize_t block_read(block_thread_state_t *block_thread_state, block_t *block, index_t *index, char *dst, size_t size);
