+++ src/main.c
@@ -547,7 +547,7 @@ int do_mount(int argc, char *argv[], int idx) {
 	char* arg0_a = argv[0];
 	const char* arg0_b = " mount [...] --";
 	size_t arglen = strlen(arg0_a) + strlen(arg0_b) + 1;
-	for (size_t i = optind; i < argc; i++) {
+	for (int i = optind; i < argc; i++) {
 		arglen += strlen(argv[i]) + 1;
 	}
 	const int fuse_argc = argc - optind + 1;
@@ -566,7 +566,7 @@ int do_mount(int argc, char *argv[], int idx) {
 	memcpy(argpos, arg0_a, strlen(arg0_a)); argpos += strlen(arg0_a);
 	memcpy(argpos, arg0_b, strlen(arg0_b)); argpos += strlen(arg0_b);
 	*argpos = 0; argpos++;
-	for (size_t i = optind; i < argc; i++) {
+	for (int i = optind; i < argc; i++) {
 		fuse_argv[i - optind + 1] = argpos;
 		memcpy(argpos, argv[i], strlen(argv[i]));
 		argpos += strlen(argv[i]);
